+++ android/src/main/java/com/dooboolab/fluttersound/FlutterSoundPlayer.java
@@ -70,7 +70,7 @@ class FlautoPlayerPlugin
 
 	void invokeMethod ( String methodName, Map dic )
 	{
-		Log.d(TAG, "FlutterAutoPlugin: invokeMethod" + methodName);
+		//Log.d(TAG, "FlutterAutoPlugin: invokeMethod" + methodName);
 		channel.invokeMethod ( methodName, dic );
 	}
 
@@ -457,7 +457,7 @@ public class FlutterSoundPlayer implements MediaPlayer.OnErrorListener
 				// long time = mp.getCurrentPosition();
 				// DateFormat format = new SimpleDateFormat("mm:ss:SS", Locale.US);
 				// final String displayTime = format.format(time);
-				Log.d(TAG, "Running updateProgress");
+				//Log.d(TAG, "Running updateProgress");
 				try {
 					JSONObject json = new JSONObject();
 					json.put("duration", String.valueOf(mp.getDuration()));
@@ -465,7 +465,7 @@ public class FlutterSoundPlayer implements MediaPlayer.OnErrorListener
 					mainHandler.post(new Runnable() {
 						@Override
 						public void run() {
-							Log.d(TAG, "Invoking updateProgress: " + json.toString());
+							//Log.d(TAG, "Invoking updateProgress: " + json.toString());
 							invokeMethodWithString("updateProgress", json.toString());
 						}
 					});