+++ ios/Classes/FlutterSoundRecorder.m
@@ -155,4 +155,3 @@ extern void FlautoRecorderReg(NSObject<FlutterPluginRegistrar>* registrar)
 
 @end
 //---------------------------------------------------------------------------------------------
-